+++ extensions/docsearch/src/algolia/types.ts
@@ -1,9 +1,10 @@
 export interface IAPIData {
   name: string;
   icon: string;
-  apiKey: string;
   appId: string;
+  apiKey: string;
   indexName: string;
+  homepage: string;
   subtitle?: string;
   searchParameters?: object;
 }